refactored models into models directory for cideries and flavour profiles.
